+++ package.json
@@ -121,9 +121,9 @@
     "test-types": "node ts-test/index.js && tsc --esModuleInterop true --noEmit true --strictNullChecks true --noImplicitAny true --strict true ts-test/*.ts",
     "eslint": "eslint '**/*.{js,md,ts}' --max-warnings 0 --ignore-path ./.eslintignore",
     "eslint-fix": "npx eslint --fix '**/*.{js,md,ts}' --max-warnings 0 --ignore-path ./.eslintignore",
-    "test": "NODE_ENV=test mocha --exit --bail --timeout 15000 --require ./babel-register test/*.js --async-stack-traces",
+    "test": "NODE_ENV=test mocha --exit --bail --timeout 15000 --require ./babel-register test/integration/*.js --async-stack-traces",
     "test-local": "STREAM_LOCAL_TEST_RUN=true yarn test",
-    "testall": "NODE_ENV=test mocha --exit --timeout 3000 --require ./babel-register test/*.js --async-stack-traces",
+    "testall": "NODE_ENV=test mocha --exit --timeout 3000 --require ./babel-register test/integration/*.js --async-stack-traces",
     "testwatch": "NODE_ENV=test nodemon ./node_modules/.bin/mocha --timeout 15000 --require test-entry.js test/test.js",
     "lint": "yarn run prettier && yarn run eslint",
     "lint-fix": "yarn run prettier-fix && yarn run eslint-fix",
