+++ test/integration/channels.js
