+++ test/integration/custom_commands.js
