+++ test/integration/delete_channel.js
