+++ test/integration/hooks.js
