+++ test/integration/silent.js
