+++ test/integration/translate.js
