+++ test/integration/utils.js
@@ -1,4 +1,4 @@
-import { StreamChat } from '../src';
+import { StreamChat } from '../../src';
 import chai from 'chai';
 const expect = chai.expect;
 require('dotenv').config();
